Add files via upload
So I decided to experiment with nationalist attraction to try and address the crazy nationalist rates in Africa.

I noticed that there was no negative attraction to Nationalist for accepted cultures, only primary cultures, so even if you passed the Nation of Equals Act African countries would still have up to 60% nationalist support in their provinces. I decided to make it so that there was a -300% attraction to nationalist for both primary and accepted cultures if the country was not a Nationalist dictatorship. This way, there would be high nationalist support after independence for a few years, but it would decline over the decade to very low numbers in places with primary and accepted pops. 

I also noticed that Nationalists would get an attraction bonus from high militancy and consciousness even when Nationalists were in power. I changed this so that this bonus would only be present if nationalists were not in power, so you didn't get strange situations where nationalists refused